feat: User enum 위치 변경
